+++ app/lecture/fetch.tsx
@@ -10,7 +10,42 @@ import { DownIcon } from '@/component/icon';
 import { Grid } from './aux';
 import { ELEM_PER_PAGE } from '@/lib/directive';
 import { useApi } from '@/lib/hooks/api';
-import SearchResults from './results';
+
+import { memo } from 'react';
+import { apiSearch } from '@/lib/api-client/calls/course';
+
+import CoursePreview from '@/component/view/CoursePreview';
+import { Box, CourseLoadingItems } from './aux';
+
+const SearchResults = memo(
+  function SearchResults({ keyword, order, page }: ReqSearchCourse) {
+    const { loading, response } = useApi(apiSearch, { keyword, order, page });
+
+    if (loading) {
+      return <CourseLoadingItems />;
+    }
+
+    if (response.status !== 'SUCCESS') {
+      return <Box>오류가 발생했습니다.</Box>;
+    }
+
+    return (
+      <>
+        {response.data.map((course) => (
+          <div key={course.course_id} className="animate-fade">
+            <CoursePreview key={course.course_id} course={course} />
+          </div>
+        ))}
+      </>
+    );
+  },
+  (prev, next) =>
+    prev != undefined &&
+    next != undefined &&
+    prev.keyword === next.keyword &&
+    prev.order === next.order &&
+    prev.page === next.page,
+);
 
 function SearchResultsViewWithOrder({
   query: keyword,
